+++ lib/phoenix_kit_web/integration.ex
@@ -152,7 +152,7 @@ defmodule PhoenixKitWeb.Integration do
           live "/admin/users/referral-codes/edit/:id", Live.Users.ReferralCodeFormLive, :edit
           live "/admin/emails/dashboard", Live.EmailTracking.EmailMetricsLive, :index
           live "/admin/emails", Live.EmailTracking.EmailLogsLive, :index
-          live "/admin/emails/:id", Live.EmailTracking.EmailDetailsLive, :show
+          live "/admin/emails/email/:id", Live.EmailTracking.EmailDetailsLive, :show
           live "/admin/emails/queue", Live.EmailTracking.EmailQueueLive, :index
           live "/admin/emails/blocklist", Live.EmailTracking.EmailBlocklistLive, :index
         end